migrate to /v2 dir
