--- app/helpers/bootstrap_rails_helpers/glyph_helper.rb
@@ -1,11 +1,0 @@
-module BootstrapRailsHelpers::GlyphHelper
-  # ==== Examples
-  # glyph(:share_alt)
-  # # => <i class="icon-share-alt"></i>
-  # glyph(:lock, :white)
-  # # => <i class="icon-lock icon-white"></i>
-
-  def glyph(*names)
-    content_tag :i, nil, :class => names.map{|name| "icon-#{name.to_s.gsub('_','-')}" }
-  end
-end
