+++ app/views/v2/bootstrap_rails_helpers/_modal_body.html.erb
