+++ app/views/v2/bootstrap_rails_helpers/_modal_footer.html.erb
