+++ app/views/v2/bootstrap_rails_helpers/_modal_header.html.erb
