+++ lib/bootstrap_rails_helpers/railtie.rb
@@ -1,17 +1,28 @@
-require File.dirname(__FILE__) + '/../../app/helpers/bootstrap_rails_helpers/breadcrumbs_helper'
-require File.dirname(__FILE__) + '/../../app/helpers/bootstrap_rails_helpers/flash_helper'
-require File.dirname(__FILE__) + '/../../app/helpers/bootstrap_rails_helpers/glyph_helper'
-require File.dirname(__FILE__) + '/../../app/helpers/bootstrap_rails_helpers/modal_helper'
+require 'bootstrap_rails_helpers/v2/breadcrumbs_helper'
+require 'bootstrap_rails_helpers/v2/flash_helper'
+require 'bootstrap_rails_helpers/v2/glyph_helper'
+require 'bootstrap_rails_helpers/v2/modal_helper'
 
 module BootstrapRailsHelpers
   class Railtie < Rails::Railtie
-    initializer "bootstrap_rails_helpers.breadcrumbs_helper" do
-      ActionView::Base.send :include, BootstrapRailsHelpers::BreadcrumbsHelper
-      ActionView::Base.send :include, BootstrapRailsHelpers::FlashHelper
-      ActionView::Base.send :include, BootstrapRailsHelpers::GlyphHelper
-      ActionView::Base.send :include, BootstrapRailsHelpers::ModalHelper
-      ActionController::Base.append_view_path File.dirname(__FILE__) + "/../../app/views"
+    config.after_initialize do
+      if BootstrapRailsHelpers.bootstrap_version == 3
+        # TODO: load bootstrap 3
+      elsif BootstrapRailsHelpers.bootstrap_version == 2
+        ActiveSupport.on_load :action_view do
+          include BootstrapRailsHelpers::V2::BreadcrumbsHelper
+          include BootstrapRailsHelpers::V2::FlashHelper
+          include BootstrapRailsHelpers::V2::GlyphHelper
+          include BootstrapRailsHelpers::V2::ModalHelper
+        end
+        ActiveSupport.on_load :action_controller do
+          append_view_path File.dirname(__FILE__) + "/../../app/views/v2"
+        end
+      else
+        # TODO: raise error
+      end
     end
+
   end
 end
