+++ lib/bootstrap_rails_helpers/v2/glyph_helper.rb
@@ -1,0 +1,15 @@
+module BootstrapRailsHelpers
+  module V2
+    module GlyphHelper
+      # ==== Examples
+      # glyph(:share_alt)
+      # # => <i class="icon-share-alt"></i>
+      # glyph(:lock, :white)
+      # # => <i class="icon-lock icon-white"></i>
+
+      def glyph(*names)
+        content_tag :i, nil, :class => names.map{|name| "icon-#{name.to_s.gsub('_','-')}" }
+      end
+    end
+  end
+end
