+++ spec/lib/bootstrap_rails_helpers/v2/breadcrumbs_helper_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe BootstrapRailsHelpers::BreadcrumbsHelper do
+describe BootstrapRailsHelpers::V2::BreadcrumbsHelper do
 
   context 'with dummy class' do
     class DummyClass; end
     before(:all) do
-      DummyClass.send :include, BootstrapRailsHelpers::BreadcrumbsHelper
+      DummyClass.send :include, BootstrapRailsHelpers::V2::BreadcrumbsHelper
       @dummy = DummyClass.new
     end
     it { @dummy.should respond_to(:add_breadcrumb) }
