+++ spec/lib/bootstrap_rails_helpers/v2/flash_helper_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe BootstrapRailsHelpers::FlashHelper do
+describe BootstrapRailsHelpers::V2::FlashHelper do
 
   context 'with dummy class' do
     class DummyClass; end
     before(:all) do
-      DummyClass.send :include, BootstrapRailsHelpers::FlashHelper
+      DummyClass.send :include, BootstrapRailsHelpers::V2::FlashHelper
       @dummy = DummyClass.new
     end
     it { @dummy.should respond_to(:alert_message) }
