+++ spec/lib/bootstrap_rails_helpers/v2/glyph_helper_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe BootstrapRailsHelpers::GlyphHelper do
+describe BootstrapRailsHelpers::V2::GlyphHelper do
 
   context 'with dummy class' do
     class DummyClass; end
     before(:all) do
-      DummyClass.send :include, BootstrapRailsHelpers::GlyphHelper
+      DummyClass.send :include, BootstrapRailsHelpers::V2::GlyphHelper
       @dummy = DummyClass.new
     end
     it { @dummy.should respond_to(:glyph) }
